refactor: setting dialog
